libs/aries-basic-controller/requirements.txt
aiohttp~=3.7.3
asyncio
prompt_toolkit
pygments
PyPubSub
python-dotenv
termcolor
pillow
qrcode
beautifulsoup4
dataclasses